Move test util functions to 'testutils' package
