+++ backend/testutils/http.go
@@ -1,4 +1,4 @@
-package testing
+package testutils
 
 import (
 	"io"
